Most of the way through a rewrite of the Config system
The idea here is to replace the annoying XML config file with just doing
it in code, which seems to be a bit more idiomatic for Scala.  Turns out
to work nicely.  The new class Config2 will replace the old Config -- it
declares the world inline directly as a data structure.

Had lots of fun making this more and more concise, and you will see that
Config2 gets even moreso mid-file, as I figured out that I should use an
implicit to reduce yet more boilerplate.  And throughout, I'm using the
Scala convention of object.apply() to concisely build the related class.

The end result isn't *quite* as concise as I might wish (because I
haven't found good syntactic sugar to replace the Seq declarations), but
it's pretty close: Config2 is pleasantly readable.

Note that this isn't finished yet.  It's producing additional errors
because I haven't dealt with Barons and Champions yet, and then I need
to switch the actual file-processing over.  But it's a good start, and a
major improvement.