Added Lucan III
Major enhancement: added the ability to parse the "old Word style" Court
Reports.  These look similar to the modern ones, but their captions are
simple paragraphs before the table, so had to add additional machinery
to parse that.

Added the ability to cope with abbreviated month names.

Fixed a bug in the court report parser: we weren't fully scrubbing cell
contents.