Fixed the Fencing Orders, and added the TDL
Realized that I had little confidence in the way I'd assigned the
various "Fence" names to the Guardsman and OGR, so did some hard
correlation between their award lists and the court reports.  I think
the results are now consistent and correct.

Also realized that I'm in danger of losing track of everything that
needs to be done, so wrote up a ToDoList for the remainder of the
project.