Added Siegfried I
Note that Siegfried1.htm is a heavily hand-edited version of
Siegfried1.doc, to make it parse properly in the Ancient parser.

Also tweaked the Ancient parser to cope with em dashes.  Took ages to
figure out that that was why it wasn't working.

Also introduced the ability to simply put commentary business in square
brackets to mark it.  Should have done that ages ago.  Fixed up all of
the places where I had used the clumsier "BUSINESS:" marker before.