Added Vis I
Also opened a can of worms: realized that I wasn't always stripping
asterisks off of names.  When I did so, found a number of other cases
that weren't parsing correctly.

Added support for plain parens in a name entry to serve as an AKA, with
no explicit reference type.
Added support for dashes as a separator in a reference.
Fixed a problem where whitespace in the middle of a name line, between
the boldfaces, would get completely scrubbed away, producing nonsense.
Stopped stripping 160's entirely; instead treat them as whitespace.
Fixed several problematic entries in the alpha list.