More parenthetical fixes
Refactored the parenthetical handling from the AlphaParser into
OPFileParser, so that CourtReportParser can use it.  This is useful, and
fixes a bunch of errors, although it introduces all sorts of new edge
cases.

Also, we now cope with "(of Floobshire)" at the end of names, and handle
that more or less properly.

Fixed a place that needs to scrub persona names.