names.conf now includes things we got from the OP
Dedup'er is only visible in the comments, but we do include everything
we have sussed from explicit cross-references in the OP.

Weakened the cross-referencing to *not* include parentheticals from
court reports, which are frequently not names.

Fixed several blatant typos in the alpha listing, which were kicked up
in this process.

Fixed the fact that the name config loader didn't cope with UTF8 input.