Fixed the Great Magnus / Marcus Disaster
It's pretty clear that, at some point, somebody did a semi-global search
and replace of the name "Magnus" for "Marcus" -- we have a pile of
Marcuses where the Magnuses belong.

My best guess is that Marcus Athienou's Salamander's Tear got recorded
as Magnus Athienou instead, and when the error was pointed out, instead
of moving the award to the existing record for Marcus, the name "Magnus"
just got replaced with "Marcus".  Worse, this was apparently done as a
global, because every instead of "Magnus" after him was *also* changed
to "Marcus", causing a number of bogusly named records.

I *think* I have fixed all the errors, but it's hard to be sure.